Don't try to cleanup non-existing ios Simulators https://bugs.webkit.org/show_bug.cgi?id=158510
Reviewed by Alexey Proskuryakov.

* Scripts/webkitpy/port/ios.py:
(IOSSimulatorPort.clean_up_test_run):
Check if simulator exists before trying to clean it up.


Canonical link: https://commits.webkit.org/176595@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201812 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/Scripts/webkitpy/port/ios.py
@@ -254,6 +254,8 @@ class IOSSimulatorPort(Port):
                 pass
 
         for i in xrange(self.child_processes()):
+            if not os.path.exists(self.get_simulator_path(i)):
+                continue
             try:
                 subprocess.call([self.LSREGISTER_PATH, "-v", "-u", self.get_simulator_path(i)])
                 shutil.rmtree(self.get_simulator_path(i), ignore_errors=True)
